[week4] 4주차 앱 배포

+++ app/src/main/java/com/ariari/mowoori/ui/stamp/entity/Stamp.kt
@@ -12,5 +12,5 @@ data class Stamp(
 data class StampInfo(
     val pictureUrl: String = "",
     val comment: String = "",
-    val timeStamp: Long = 0,
+    val timeStamp: Int = 0,
 ) : Parcelable
